feat: hide Create Alert CTA in the header
