+++ src/app/Scenes/Artwork/Components/ArtworkScreenHeaderCreateAlert.tsx
@@ -1,6 +1,9 @@
 import { BellIcon, Button } from "@artsy/palette-mobile"
 import { ArtworkScreenHeaderCreateAlert_artwork$key } from "__generated__/ArtworkScreenHeaderCreateAlert_artwork.graphql"
 import { CreateArtworkAlertModal } from "app/Components/Artist/ArtistArtworks/CreateArtworkAlertModal"
+import { hasBiddingEnded } from "app/Scenes/Artwork/utils/hasBiddingEnded"
+import { isLotClosed } from "app/Scenes/Artwork/utils/isLotClosed"
+import { useFeatureFlag } from "app/utils/hooks/useFeatureFlag"
 import { useState } from "react"
 import { graphql, useFragment } from "react-relay"
 
@@ -16,10 +19,14 @@ export const ArtworkScreenHeaderCreateAlert: React.FC<ArtworkScreenHeaderCreateA
     artworkRef
   )
   const [showCreateArtworkAlertModal, setShowCreateArtworkAlertModal] = useState(false)
-  const { isForSale } = artwork
+  const { isForSale, sale, saleArtwork } = artwork
   const hasArtists = !!artwork?.artists?.length
 
-  if (!hasArtists) {
+  const isLotClosedOrBiddingEnded =
+    hasBiddingEnded(sale, saleArtwork) || isLotClosed(sale, saleArtwork)
+  const enableAuctionHeaderAlertCTA = useFeatureFlag("AREnableAuctionHeaderAlertCTA")
+
+  if (!hasArtists || (isLotClosedOrBiddingEnded && enableAuctionHeaderAlertCTA)) {
     return null
   }
 
@@ -51,5 +58,14 @@ const ArtworkScreenHeaderCreateAlert_artwork = graphql`
       internalID
     }
     isForSale
+    sale {
+      isClosed
+      startAt
+    }
+    saleArtwork {
+      endAt
+      endedAt
+      extendedBiddingEndAt
+    }
   }
 `
