+++ src/app/Scenes/Artwork/helpers.ts
@@ -1,8 +1,8 @@
 import { useSpace } from "@artsy/palette-mobile"
 import { isPad } from "app/utils/hardware"
+import { ScreenDimensionsWithSafeAreas, useScreenDimensions } from "app/utils/hooks"
 import { findRelayRecord, findRelayRecordByDataID } from "app/utils/relayHelpers"
 import { Record } from "relay-runtime/lib/store/RelayStoreTypes"
-import { ScreenDimensionsWithSafeAreas, useScreenDimensions } from "app/utils/hooks"
 import { getMeasurements } from "./Components/ImageCarousel/geometry"
 
 export const getDefaultImageDimensions = (
